small edit

+++ Virtual Machine/client/checksum.c
@@ -1,6 +1,6 @@
 #include "checksum.h"
 
-uLong computeChecksum(char* buffer, size_t length)
+uLong computeChecksum(char* buffer, long length)
 {
 	uLong crc = crc32(0L, Z_NULL, 0);
